Move createSelectedNvts
